moved all stored procedure handling code (except get_procedures_for_schema) to SQLAlchemySource
